32-16 bit current
why were you even using 32-bit variables for 12 bits of data transmitted as a signed 16-bit packet?

+++ src/INA3221.cpp
@@ -396,7 +396,7 @@ void INA3221::setCurrentSumDisable(ina3221_ch_t channel) {
 }
 
 int32_t INA3221::getShuntVoltage(ina3221_ch_t channel) {
-    int32_t res;
+    int16_t res;
     ina3221_reg_t reg;
     uint16_t val_raw = 0;
 
@@ -416,7 +416,7 @@ int32_t INA3221::getShuntVoltage(ina3221_ch_t channel) {
 
     // instead of bit-shifting, (which would break the signed integer signing,) divide by 8 to remove the (reserved) last 3 least-significant bits
     // 1 LSB = 40uV
-    res = (int32_t)(val_raw);
+    res = (int16_t)(val_raw);
     res = (res / 8) * 40;
 
     return res;
